Rename contact-list-item.tsx to item.tsx
